requirements.txt
# Core dependencies
numpy>=1.23.5
pillow>=9.5.0
opencv-python-headless>=4.7.0.72
streamlit>=1.22.0

# OCR and ML dependencies
python-doctr[torch]>=0.6.0
pytesseract>=0.3.10
transformers
datasets
seqeval
pyarrow>=14.0.1

# Development dependencies
python-dotenv>=1.0.0
pyyaml>=6.0.1


#
pdf2image>=1.16.3
PyMuPDF>=1.19.6
plotly>=5.13.0
python-Levenshtein>=0.12.2
fuzzywuzzy>=0.18.0